requirements.txt
pandas==2.0.3
python_dateutil==2.8.2
selenium==4.23.1
webdriver_manager==4.0.2
